[TASK] Add Projektplan as Gantt Diagram
OpenProject is currently running on my local docker mashine. It is a really nice tool for project management, but I am thinking about hosting it on maispace.de as a subdomain... maybe later in the project process. Currently I want to use the Github Features for a maximum of collaboration with my mentor and for transparency...